requirements.txt
    # local package
-e .

numpy
pandas
matplotlib
seaborn
librosa
IPython
dataprep
plotly
soundfile
sklearn
tqdm

torch
torchaudio
torchmetrics
tensorflow
keras
mlflow
streamlit

pyAudioAnalysis
scipy

bs4
requests

imblearn
hmmlearn
pydub
eyed3
boto

# external requirements
click~=8.0.4
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1


# backwards compatibility
pathlib2

setuptools~=58.0.4